【代码计数:功能】完成不同类型文件支持
