+++ WIN7U-20170923L.ccl
@@ -1,3 +1,1 @@
-2018,6,16,2,0,7,02018,6,19,-1,1,3,0
-2018,6,23,1,1,4,0
-2018,6,24,15,16,118,0
+2018,7,19,169,66,832,0,py2018,7,19,1424,0,7094,0,html2018,7,19,0,0,921,0,xml
