+++ config.ini
@@ -1,0 +1,8 @@
+[sec_a]
+supported_suffix=.py,.java,.html,.js,.xml,.css
+py=#
+java=//
+html=<!--
+js=//
+xml=<!--
+css=<!--
