+++ file_op.py
@@ -1,94 +1,87 @@
 import chardet
 import os
 import codecs
+import socket
 
 
-def count_code(path):
+def count_code(path, suffix):
     i = 0   # 递归迭代函数中，函数级变量代码定义的位置不对，实现逻辑就会出错
     path_list = os.listdir(path)
     for filename in path_list:
         if os.path.isfile(path + '/' + filename):
-            if filename.endswith('.py'):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
-                encoding =  get_encoding(path + '/' + filename)
+            if filename.endswith(suffix):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
+                encoding = get_encoding(path + '/' + filename)
                 if encoding is None:
                     encoding = 'utf-8'
-                #print(filename + ' ' + encoding)
-                f = codecs.open(path + '/' + filename, encoding = encoding)
+                f = codecs.open(path + '/' + filename, encoding=encoding)
                 k = 0
                 for line in f.readlines():
                     line = line.strip()
                     if not line.startswith('#') and len(line):
                         k += 1
                 i += k
-                #print(str(k) + '行')
 
         if os.path.isdir(path + '/' + filename):
             current_path = path + '/' + filename
-            j = count_code(current_path)
+            j = count_code(current_path, suffix)
             if j:
                 i = i + j
     return i    # 递归迭代函数中，return 用于从函数内部传递返回值出来；位置决定递归函数的返回结束！
 
 
-def count_newline(path):
+def count_newline(path, suffix):
     i = 0   # 递归迭代函数中，函数级变量代码定义的位置不对，实现逻辑就会出错
     path_list = os.listdir(path)
 
     for filename in path_list:
         if os.path.isfile(path + '/' + filename):
-            if filename.endswith('.py'):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
-                #print(filename + ' ' + get_encoding(path + '/' + filename))
-                f = codecs.open(path + '/' + filename, encoding = get_encoding(path + '/' + filename))
+            if filename.endswith(suffix):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
+                f = codecs.open(path + '/' + filename, encoding=get_encoding(path + '/' + filename))
                 k = 0
                 for line in f.readlines():
                     line = line.strip()
                     if not len(line):
                         k += 1
                 i += k
-                #print(str(k) + '行')
         if os.path.isdir(path + '/' + filename):
             current_path = path + '/' + filename
-            j = count_newline(current_path)
+            j = count_newline(current_path, suffix)
             if j:
                 i = i + j
     return i    # 递归迭代函数中，return 用于从函数内部传递返回值出来；位置决定递归函数的返回结束！
 
 
-def count_comment(path):
+def count_comment(path, suffix, char_comment):
     i = 0   # 递归迭代函数中，函数级变量代码定义的位置不对，实现逻辑就会出错
     path_list = os.listdir(path)
 
     for filename in path_list:
         if os.path.isfile(path + '/' + filename):
-            if filename.endswith('.py'):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
-                #print(filename + ' ' + get_encoding(path + '/' + filename))
-                f = codecs.open(path + '/' + filename, encoding = get_encoding(path + '/' + filename))
+            if filename.endswith(suffix):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
+                f = codecs.open(path + '/' + filename, encoding=get_encoding(path + '/' + filename))
                 k = 0
                 for line in f.readlines():
                     line = line.strip()
-                    if line.startswith('#'):
+                    if line.startswith(char_comment):
                         k += 1
                 i += k
 
-                #print(str(k) + '行')
         if os.path.isdir(path + '/' + filename):
             current_path = path + '/' + filename
-            j = count_comment(current_path)
+            j = count_comment(current_path, suffix, char_comment)
             if j:
                 i = i + j
     return i    # 递归迭代函数中，return 用于从函数内部传递返回值出来；位置决定递归函数的返回结束！
 
 
-def count_all(path):
+def count_all(path, suffix):
     i = 0   # 递归迭代函数中，函数级变量代码定义的位置不对，实现逻辑就会出错
     path_list = os.listdir(path)
     for filename in path_list:
         if os.path.isfile(path + '/' + filename):
-            if filename.endswith('.py'):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
-                #print(filename + ' ' + get_encoding(path + '/' + filename))
-                f = codecs.open(path + '/' + filename, encoding = get_encoding(path + '/' + filename))
+            if filename.endswith(suffix):    # 用 endswith() 代替 in ，精确匹配，解决读取到 .pyc 文件时的编码报错问题
+                f = codecs.open(path + '/' + filename, encoding=get_encoding(path + '/' + filename))
                 k = len(f.readlines())
-                #print(str(k) + '行')
                 i += k
         if os.path.isdir(path + '/' + filename):
             current_path = path + '/' + filename
@@ -98,18 +91,21 @@ def count_all(path):
 
     return i    # 递归迭代函数中，return 用于从函数内部传递返回值出来；位置决定递归函数的返回结束！
 
+
 # bug:如果文件为空, 返回时NoneType
 def get_encoding(file):
     with open(file, 'rb') as f:
         return chardet.detect(f.read())['encoding']
 
 
-def find_last_line_index(lines):
+def find_last_line_index(lines, suffix):
     for index in range(len(lines)-1, -1, -1):
         if len(lines[index].strip()) == 0:
             continue
         else:
-            return index
+            if lines[index].split(',')[7].find(suffix) >= 0:
+                return index
+
 
 def count_code_line_txt(path):
     code_sum = 0
@@ -126,3 +122,12 @@ def count_code_line_txt(path):
                 code_sum += int(fl[index].split(',')[5])
                 print('code_sum:' + code_sum)
     return code_sum
+
+
+def get_host_name():
+    hostname = socket.gethostname()
+    if '.local' in hostname:
+        print("mac会出现两个hostname，一个正确的，还有一个加local的。修复：去掉local")
+        hostname = hostname.replace('.local', '')
+    print('程序员主机名:' + hostname)
+    return hostname
